producer takes vector of bytes again
